Add function removeBook
This function will add an event listener to each of the buttons in the cards. It will then use the data attribute to compare the id values to make sure the card that is deleted is the one that had its button clicked. It also removes the book from the library array.